automatic scaling of the axes Primitively done
when scaled upwards, the axes begin to adjust to the new data and as such can now show extremely large data (>100). I am working on the other side of the spectrum, extremely small data (<100). Need to fix the generateTick per value, at small values it simply is not generating enough of them. It can be assumed that at large values it is generating an amount not necessary for successful program operation.