Issue with text switching fluidly.
Text switching is quite discrete and stiff. Attempting to switch two adjacent text boxes by slowly incrementing them in opposite directions but it seems an n amount of variables will need to be created. Instead, an array is used as a common construct, but resetting the array is difficult and conflicting matches can interfere with proper text switching.