adjustable x-ticks!
x-ticks are now adjustable, and images must be in the data folder and lowercase (and png). If all these conditions are met, the image will show up next to the corresponding text. Leaderboard has been improved and now the unit counter is accurate. Will apply unix epoch converter soon to turn it into human readable time.