Dataset collector now fills up a text file with bitcoin values (filtered) and <0 fixed!
The datasetCollector.py now filters any NaN values and after 630k, every 120th value, of the bitcoin csv file. This is to lower the amount of data samples so that smoothing can occur when data splicing. The graph can now recognize when values are below 0 and thus not draw them. dataSplicer.py can now add in -1 for any splined y values.