Naming system works (no animation yet)
Replaced parallel arrays with an object construct that holds all of the data for that particular frame. This allows sorting the object by its y values which also sorts the rest of the items such as hex codes, names, and image. Objects are constantly rewritten after every frame, so a better idea may be to use a setter to keep the object as is, but simply change the values and fields inside the object, and then resort.