Porting to PGraphics
Attempting to convert most of the methods in CartesianPlane to be compatible with PGraphics. This will allow saving of the image to a PImage data type which can be scaled or moved around depending on where it needs to be. A new folder may need to be created to avoid overwriting old files.