Smooth animation at beginning
Animation is now smooth at the beginning.
Fix text and image (make bigger) and adjust the autoscaling when close to boundary. Also (long-term) add a "zoom" feature to show tick marks when lines are close to each other and scaleY is too small for more refined features to be shown.