Animation complete
Animation for the grid creation is complete with each gridline being associated with another DoubleLine object. A potential problem occurs when a DoubleLine object is not created at the origin. Reflecting a point over the origin to get the other point is trivial; (x,y) --> (-x,-y). Doing so with a point that is not the origin is more difficult.
The for loop in CartesianPlane.java also needs rework.