Fixed fading text bug
Gridline drawing is now smooth and animated. The first scene of the differentiating UC video is now complete. The circle is animated and a tangent line is also animated. What seems to be left is text and narration, which a traditional video editor are much more suited to than vanim.